Navbar fixed on mobile
There was a problem with navbar that was disappearing whenever scrolled
down on mobile on Chrome browser. It seemed to look ok in Firefox and
their dev tools. After more investigation it turned out that on the
mobile view in Chrome there is an overflow and the whole screen does
not fit in the window. In the process of debugging (removing sections,
removing different CSS properties etc.) it turned out that some sections
cause the problem. Some items were to big or had additional wrappers or
paddings that caused the overflow. Items fixed: the price name for the
last item in 'Pricing' was too big; the download icons and the
descriptions had too much padding; the form in the 'Contact' sections
was wrapped in an additional div with the class of row, which stretched
it more; too many elements on the footer.